refactor: 첨부 계약을 board/common 으로 추출한다
