+++ src/main/java/inha/gdgoc/domain/board/event/dto/response/EventBoardDetailResponse.java
@@ -1,6 +1,6 @@
 package inha.gdgoc.domain.board.event.dto.response;
 
-import inha.gdgoc.domain.board.common.enums.AttachmentKind;
+import inha.gdgoc.domain.board.common.dto.AttachmentResponse;
 import inha.gdgoc.domain.board.event.enums.EventBoardStatus;
 import inha.gdgoc.domain.user.enums.TeamType;
 import java.time.Instant;
@@ -20,14 +20,4 @@ public record EventBoardDetailResponse(
     EventBoardStatus status,
     List<AttachmentResponse> attachments,
     Instant createdAt,
-    Instant updatedAt) {
-
-  public record AttachmentResponse(
-      Long id,
-      AttachmentKind kind,
-      String fileKey,
-      String fileUrl,
-      String fileName,
-      Long fileSize,
-      String url) {}
-}
+    Instant updatedAt) {}
