+++ src/main/java/inha/gdgoc/domain/board/event/service/EventBoardService.java
@@ -1,14 +1,12 @@
 package inha.gdgoc.domain.board.event.service;
 
-import inha.gdgoc.domain.board.common.entity.BoardAttachment;
-import inha.gdgoc.domain.board.common.enums.AttachmentKind;
+import inha.gdgoc.domain.board.common.dto.AttachmentResponse;
 import inha.gdgoc.domain.board.common.enums.SearchType;
+import inha.gdgoc.domain.board.common.service.AttachmentPolicy;
 import inha.gdgoc.domain.board.event.dto.request.EventBoardCreateRequest;
-import inha.gdgoc.domain.board.event.dto.request.EventBoardCreateRequest.AttachmentEntry;
 import inha.gdgoc.domain.board.event.dto.request.EventBoardUpdateRequest;
 import inha.gdgoc.domain.board.event.dto.response.DeletedEventBoardSummaryResponse;
 import inha.gdgoc.domain.board.event.dto.response.EventBoardDetailResponse;
-import inha.gdgoc.domain.board.event.dto.response.EventBoardDetailResponse.AttachmentResponse;
 import inha.gdgoc.domain.board.event.dto.response.EventBoardSummaryResponse;
 import inha.gdgoc.domain.board.event.entity.EventBoard;
 import inha.gdgoc.domain.board.event.enums.EventBoardStatus;
@@ -20,7 +18,6 @@ import inha.gdgoc.domain.user.enums.UserRole;
 import inha.gdgoc.domain.user.repository.UserRepository;
 import inha.gdgoc.global.exception.BusinessException;
 import inha.gdgoc.global.exception.GlobalErrorCode;
-import java.util.Comparator;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -33,11 +30,10 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class EventBoardService {
 
-  private static final int MAX_ATTACHMENTS = 10;
-
   private final EventBoardRepository eventBoardRepository;
   private final UserRepository userRepository;
   private final S3Service s3Service;
+  private final AttachmentPolicy attachmentPolicy;
 
   public Page<EventBoardSummaryResponse> listEventBoards(
       int page, int size, SearchType searchType, String keyword, TeamType userTeam, UserRole userRole) {
@@ -76,7 +72,7 @@ public class EventBoardService {
             authorId,
             author.getName());
 
-    applyAttachments(board, req.attachments());
+    attachmentPolicy.apply(board, req.attachments());
 
     return eventBoardRepository.save(board).getId();
   }
@@ -106,7 +102,7 @@ public class EventBoardService {
 
     if (req.attachments() != null) {
       board.getAttachments().clear();
-      applyAttachments(board, req.attachments());
+      attachmentPolicy.apply(board, req.attachments());
     }
   }
 
@@ -130,31 +126,6 @@ public class EventBoardService {
 
     requireTeamAccess(board, userRole, userTeam);
     board.restore();
-  }
-
-  private void applyAttachments(EventBoard board, List<AttachmentEntry> entries) {
-    if (entries == null) return;
-
-    if (entries.size() > MAX_ATTACHMENTS) {
-      throw new BusinessException(
-          GlobalErrorCode.BAD_REQUEST, "첨부는 최대 " + MAX_ATTACHMENTS + "개까지 등록할 수 있습니다.");
-    }
-
-    for (int i = 0; i < entries.size(); i++) {
-      AttachmentEntry entry = entries.get(i);
-
-      if (entry.url() != null && !entry.url().isBlank()) {
-        board.addLinkAttachment(entry.url(), i);
-        continue;
-      }
-
-      Long size = s3Service.getObjectSize(entry.fileKey());
-      if (size == null) {
-        throw new BusinessException(
-            GlobalErrorCode.BAD_REQUEST, "업로드되지 않은 파일입니다: " + entry.fileName());
-      }
-      board.addFileAttachment(entry.fileKey(), entry.fileName(), size, i);
-    }
   }
 
   private EventBoard findVisibleBoard(Long id, TeamType userTeam, UserRole userRole) {
@@ -198,23 +169,7 @@ public class EventBoardService {
     String thumbnailUrl =
         board.getThumbnailKey() != null ? s3Service.getS3FileUrl(board.getThumbnailKey()) : null;
 
-    List<AttachmentResponse> attachments =
-        board.getAttachments().stream()
-            .sorted(Comparator.comparingInt(BoardAttachment::getSortOrder))
-            .map(
-                a ->
-                    a.getKind() == AttachmentKind.LINK
-                        ? new AttachmentResponse(
-                            a.getId(), a.getKind(), null, null, null, null, a.getUrl())
-                        : new AttachmentResponse(
-                            a.getId(),
-                            a.getKind(),
-                            a.getFileKey(),
-                            s3Service.getS3FileUrl(a.getFileKey()),
-                            a.getFileName(),
-                            a.getFileSize(),
-                            null))
-            .toList();
+    List<AttachmentResponse> attachments = attachmentPolicy.toResponses(board.getAttachments());
 
     return new EventBoardDetailResponse(
         board.getId(),
